Fixes made from the account merger, added validation to prevent deletion of films, venues and screens when an active showing it present, CSS improvements and site flow improvements made throughout

+++ uweflix/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render
 from django.http import *
+from matplotlib.pyplot import show
 
 #Forms
 from uweflix.forms import *
@@ -15,7 +16,7 @@ from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 
 from django.contrib.auth import authenticate, login, logout
 
-# from django.contrib.auth.models import
+from django.contrib.auth.models import Group
 
 from django.views.generic import *
 
@@ -36,7 +37,9 @@ from django.contrib.auth.decorators import login_required, user_passes_test
 from django.urls import reverse
 
 def home(request):
-    return render(request, "uweflix/home.html")
+    showingList = Showing.objects.all()
+
+    return render(request, "uweflix/home.html", {'showingList': showingList})
 
 # def loginRequest(request):
 #     if request.method == 'POST':
@@ -114,9 +117,10 @@ def signupRequest(request):
             password = form.cleaned_data.get('password1')
 
             newUser = authenticate(request, username=username, password=password)
-            newUser.is_active = False
 
             newUser.save()
+            saveGroup = Group.objects.get(name='ClubRepresentative')
+            saveGroup.user_Set.add(newUser)
 
             if newUser is not None:
                 messages.success(request, 'New user account has been created {username}')
@@ -230,9 +234,27 @@ def addFilm(request):
 def deleteFilm(request, id):
 
     film = Film.objects.get(pk=id)
-    film.delete()
 
-    return redirect("allFilms")
+    try:
+        showingsOfFilm = Showing.objects.get(film_id=film) 
+    except:
+        showingsOfFilm = None
+
+    if (showingsOfFilm == None):
+        showingsOfFilmPresent = False
+    else: 
+        showingsOfFilmPresent = True
+
+    if (showingsOfFilmPresent == True):
+
+        messages.info(request, 'Error: Sorry, you cannot delete this film as it has showings, Please delete the showings first')
+
+        return render(request, "uweflix/cinemaAdmin.html")
+
+    else:
+        film.delete()
+        return redirect("allFilms")
+
 
 # Gets all films in the system to be displayed
 def getAllFilms(request):
@@ -273,9 +295,27 @@ def addVenue(request):
 def deleteVenue(request, venue_id):
 
     venue = Venue.objects.get(pk=venue_id)
-    venue.delete()
+    
+    try:
+        showingsAtVenue = Showing.objects.get(venue_id=venue) 
+    except:
+        showingsAtVenue = None
 
-    return redirect("allVenues")
+    if (showingsAtVenue == None):
+        showingsOfVenuePresent = False
+    else: 
+       showingsOfVenuePresent = True
+
+    if ( showingsOfVenuePresent == True):
+
+        messages.info(request, 'Error: Sorry, you cannot delete this venue as it has showings, Please delete the showings first')
+
+        return render(request, "uweflix/cinemaAdmin.html")
+
+    else:
+        venue.delete()
+        return redirect("allVenues")
+
 
 # Gets all venues in the system to be displayed
 def getAllVenues(request):
@@ -317,9 +357,28 @@ def addScreen(request):
 def deleteScreen(request, screen_id):
 
     screen = Screen.objects.get(pk=screen_id)
-    screen.delete()
 
-    return redirect("allScreen")
+    try:
+        showingsAtScreen= Showing.objects.get(screen_id=screen) 
+    except:
+        showingsAtScreen = None
+
+    if (showingsAtScreen == None):
+        showingsAtScreenPresent = False
+    else: 
+       showingsAtScreenPresent = True
+
+    if ( showingsAtScreenPresent == True):
+
+        messages.info(request, 'Error: Sorry, you cannot delete this screen as it has showings, Please delete the showings first')
+
+        return render(request, "uweflix/cinemaAdmin.html")
+
+    else:
+        screen.delete()
+        return redirect("allScreen")
+
+
 
 # Gets all screens in the system to be displayed
 def getAllScreens(request):
@@ -427,7 +486,6 @@ def getTicketFromShowing(request, showing_id):
     form = purchaseTicketForm(request.POST or None)
 
     showing = Showing.objects.get(pk=showing_id)
-
 
 
     if request.method == "POST":
